project_contents/requirements.txt
dlib==19.24.0
numpy==1.23.2
opencv_python_headless==4.5.5.62
Pillow==9.3.0
scipy==1.9.3
